requirements.txt
streamlit==1.35.0
streamlit-chat==0.1.1
langchain==0.2.0
langchain-community==0.2.0
faiss-cpu==1.7.4
python-dotenv==1.0.1
tqdm==4.66.4
sentence-transformers==2.7.0
transformers==4.43.2
torch>=2.0.0
requests==2.31.0
